:recycle: organize some utils
