+++ @navikt/aksel/src/codemod/transforms/v4.0.0/internal-react/internal-react.ts
@@ -1,5 +1,5 @@
 import { getLineTerminator } from "../../../utils/lineterminator";
-import moveAndRenameImport from "../../../utils/moveAndRenameImport";
+import moveAndRenameImport from "../../../utils/packageImports";
 
 /**
  * @param {import('jscodeshift').FileInfo} file
